+++ packages/expo-router/CHANGELOG.md
@@ -185,6 +185,7 @@ _This version does not introduce any user-facing changes._
 - remove error when link with preview is used for external link ([#38939](https://github.com/expo/expo/pull/38939) by [@Ubax](https://github.com/Ubax))
 - [Internal] Remove internal `SafeAreaViewSlot` to remove `SafeAreaView` import warning ([#38899](https://github.com/expo/expo/pull/38899) by [@kitten](https://github.com/kitten))
 - improve animation for stack push from different tab ([#38927](https://github.com/expo/expo/pull/38927) by [@Ubax](https://github.com/Ubax))
+- [modal] add `onDetentChange`  event ([#38143](https://github.com/expo/expo/pull/38143) by [@lodev09](https://github.com/lodev09))
 
 ### 💡 Others